1) Fixed a logic bug in earlier putback 2) Fixed Issue 352 : VC compiler generates warning on Windows AMD64 build 3) Clean up build script and code to eliminate warning from VC compiler.
git-svn-id: https://svn.java.net/svn/j3d-core~svn/trunk@714 ba19aa83-45c5-6ac9-afd3-db810772062c

+++ src/native/ogl/GeometryArrayRetained.c
@@ -163,7 +163,7 @@ executeGeometryArray(
 
     jfloat *verts, *startVertex, *clrs, *startClrs;
     jint  i;
-	size_t bstride, cbstride;
+    jint bstride, cbstride;
     jsize strip_len;
     GLsizei *strips;
     GLenum iaFormat;
@@ -598,7 +598,7 @@ JNIEXPORT
 
     jfloat *verts;
     jint i, j;
-	size_t bstride;
+    jint bstride;
     jint texStride, *texCoordSetMapOffset;
     int vAttrStride = 0;
     int vAttrOff;
@@ -1912,7 +1912,7 @@ executeIndexedGeometryArray(
     jfloat *verts,*clrs;
     jint *indices;
     jint  i;
-	size_t bstride, cbstride;
+    jint bstride, cbstride;
     jsize strip_len;
     GLsizei *countArray;
     int offset;
@@ -2982,7 +2982,7 @@ Java_javax_media_j3d_NativePipeline_buildIndexedGeometry(
     jfloat *verts;
     jint *indices;
     jint i;
-    size_t bstride;
+    jint bstride;
     jint texStride, *texCoordSetMapOffset, texSize;
     GLsizei *countArray;
     GLenum iaFormat;
